refactor(settings): extract shared search content switcher and drop imports
- Unify duplicate _SearchContentSwitcher from phone list and tablet pane
- Remove now-unused imports (settingsSearchItem / legacy search field path)
- untranslated.txt stays {} and settingsSection* reference count stays zero
