+++ lib/screens/settings/category_list_view.dart
@@ -1,14 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../l10n/l10n_ext.dart';
-import 'categories/settings_category_list.dart';
 import 'category_detail_screen.dart';
-import 'search/settings_search_item.dart';
-import 'search/settings_search_provider.dart';
-import 'search/settings_search_results_view.dart';
-import 'widgets/settings_app_bar_title.dart';
+import 'search/settings_search_content_switcher.dart';
 import 'widgets/settings_search_field.dart';
+import 'widgets/settings_app_bar_title.dart';
 
 /// スマホ（<600dp）のカテゴリ一覧画面。
 ///
@@ -46,12 +42,20 @@ class SettingsCategoryListView extends StatelessWidget {
           SliverPadding(
             padding: const EdgeInsets.fromLTRB(0, 0, 0, 100),
             sliver: SliverToBoxAdapter(
-              child: _SearchContentSwitcher(
+              child: SettingsSearchContentSwitcher(
                 onResultTap: (context, item) {
                   Navigator.of(context).push(
                     MaterialPageRoute(
                       builder: (_) =>
                           SettingsCategoryDetailScreen(category: item.category),
+                    ),
+                  );
+                },
+                onCategorySelected: (category) {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (_) =>
+                          SettingsCategoryDetailScreen(category: category),
                     ),
                   );
                 },
@@ -64,34 +68,11 @@ class SettingsCategoryListView extends StatelessWidget {
   }
 }
 
-/// クエリ非空時は検索結果ビュー、空時はカテゴリ一覧を表示する切り替え。
-class _SearchContentSwitcher extends ConsumerWidget {
-  final void Function(BuildContext context, SettingsSearchItem item)
-  onResultTap;
-
-  const _SearchContentSwitcher({required this.onResultTap});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final query = ref.watch(settingsSearchProvider).trim();
-    if (query.isNotEmpty) {
-      return SettingsSearchResultsView(onResultTap: onResultTap);
-    }
-    return SettingsCategoryList(
-      onCategorySelected: (category) {
-        Navigator.of(context).push(
-          MaterialPageRoute(
-            builder: (_) => SettingsCategoryDetailScreen(category: category),
-          ),
-        );
-      },
-    );
-  }
-}
-
 /// 検索フィールドのピン留めヘッダデリゲート。
 ///
-/// 高さはフィールド（padding 16+8+8 + 入力欄 ~48）を収める固定値。
+/// 高さは実測（TextField kMinInteractiveDimension 48 + padding 16 = 64）。
+/// min/max を等しくしないと pinned ヘッダーの paintExtent と layoutExtent が
+/// 食い違い SliverGeometry 検証に失敗するため、実高さに一致させる。
 class _SearchFieldHeaderDelegate extends SliverPersistentHeaderDelegate {
   final Widget child;
 
@@ -103,9 +84,6 @@ class _SearchFieldHeaderDelegate extends SliverPersistentHeaderDelegate {
   @override
   double get maxExtent => _kSearchHeaderHeight;
 
-  /// 高さは実測（TextField kMinInteractiveDimension 48 + padding 16 = 64）。
-  /// min/max を等しくしないと pinned ヘッダーの paintExtent と layoutExtent が
-  /// 食い違い SliverGeometry 検証に失敗するため、実高さに一致させる。
   static const double _kSearchHeaderHeight = 64;
 
   @override
